refactor methods
